+++ src/mesh/mesh.bool.tpp
@@ -80,7 +80,7 @@ private: // methods
             if(entry.data.is_isct) {
                 ShortVec<uint, 2> tid0s;
                 ShortVec<uint, 2> tid1s;
-				for(size_t ind=0; ind!=entry.tids.size(); ++ind)
+				for(uint ind=0; ind!=entry.tids.size(); ++ind)
 				{
 					const uint& tid = entry.tids[ind];
                     if(boolData(tid) & 1)
@@ -184,7 +184,7 @@ void Mesh<VertData,TriData>::BoolProblem::doSetup(
     // of the two input meshes.
     // These components are not necessarily uniformly inside or outside
     // of the other operand mesh.
-    UnionFind uf(mesh->tris.size());
+    UnionFind uf(static_cast<uint>(mesh->tris.size()));
     for_ecache([&](uint, uint, bool, const ShortVec<uint, 2> &tids) {
         uint tid0 = tids[0];
         for(uint k=1; k<tids.size(); k++)
@@ -197,10 +197,10 @@ void Mesh<VertData,TriData>::BoolProblem::doSetup(
     for(uint i=0; i<mesh->tris.size(); i++) {
         uint ufid = uf.find(i);
         if(uq_ids[ufid] == uint(-1)) { // unassigned
-            uint N = components.size();
+            size_t N = components.size();
             components.push_back(std::vector<uint>());
             
-            uq_ids[ufid] = uq_ids[i] = N;
+            uq_ids[ufid] = uq_ids[i] = static_cast<uint>(N);
             components[N].push_back(i);
         } else { // assigned already
             uq_ids[i] = uq_ids[ufid]; // propagate assignment
@@ -255,9 +255,9 @@ void Mesh<VertData,TriData>::BoolProblem::doSetup(
                 auto &entry = ecache(a,b);
                 byte inside_sig = boolData(curr_tid) & 2;
                 if(entry.data.is_isct)  inside_sig ^= 2;
-                for(size_t ind=0; ind!=entry.tids.size(); ++ind)
-				{
-					const uint& tid = entry.tids[ind];
+                for(uint ind=0; ind!=entry.tids.size(); ++ind)
+                {
+                    const uint& tid = entry.tids[ind];
                     if(visited[tid])                    continue;
                     if((boolData(tid)&1) != operand)    continue;
